requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9

numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66

rasterio>=1.3
geopandas>=0.14
shapely>=2.0
simplekml>=1.3

matplotlib>=3.7
seaborn>=0.12

Pillow>=10.0
tifffile>=2023